pyproject.toml
[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[project]
name = "sym-cps"
description = "Develop AI-based approaches to enable correct-by-construction design of Cyber Physical Systems"
authors = [{name = "LOGiCS Team UC Berkeley", email = "mallozzi@berkeley.edu"}]
license-expression = "MIT"
readme = "README.md"
keywords = []
dynamic = ["version"]
dependencies = [
    "networkx>=2.8.6",
    "dill>=0.3.5.1",
    "python-igraph>=0.10.0",
    "matplotlib>=3.5.3",
    "scipy>=1.9.1",
]
requires-python = ">=3.10,<3.11"

[project.scripts]
parse-library = "sym_cps.cli:run_parse_library"
load-library = "sym_cps.cli:run_load_library"
export-design = "sym_cps.cli:run_export_design"
load-design = "sym_cps.cli:run_load_design_json"

[tool.pdm]
version = {source = "scm"}
package-dir = "src"


[tool.pdm.build]
package-dir = "src"
editable-backend = "editables"

[tool.pdm.dev-dependencies]
duty = ["duty>=0.7"]
docs = [
    "mkdocs>=1.3",
    "mkdocs-coverage>=0.2",
    "mkdocs-gen-files>=0.3",
    "mkdocs-literate-nav>=0.4",
    "mkdocs-material>=7.3",
    "mkdocs-section-index>=0.3",
    "mkdocstrings[python]>=0.18",
    "markdown-callouts>=0.2",
    "markdown-exec>=0.5",
    "toml>=0.10",
]
format = [
    "autoflake>=1.4",
    "black>=21.10b0",
    "isort>=5.10",
]
maintain = [
    "git-changelog>=0.4",
]
quality = [
    "darglint>=1.8",
    "flake8<4",  # TODO: remove once importlib-metadata version conflict is resolved
    "flake8-bandit>=2.1",
    "flake8-black>=0.2",
    "flake8-bugbear>=21.9",
    "flake8-builtins>=1.5",
    "flake8-comprehensions>=3.7",
    "flake8-docstrings>=1.6",
    "flake8-pytest-style>=1.5",
    "flake8-string-format>=0.3",
    "flake8-tidy-imports>=4.5",
    "flake8-variables-names>=0.0",
    "pep8-naming>=0.12",
    "wps-light>=0.15",
]
tests = [
    "pytest>=6.2",
    "pytest-cov>=3.0",
    "pytest-randomly>=3.10",
    "pytest-xdist>=2.4",
]
typing = [
    "mypy>=0.910",
    "types-markdown>=3.3",
    "types-toml>=0.10",
]
security = ["safety>=2"]

[tool.black]
line-length = 120
exclude = "tests/fixtures"

[tool.isort]
line_length = 120
not_skip = "__init__.py"
multi_line_output = 3
force_single_line = false
balanced_wrapping = true
default_section = "THIRDPARTY"
known_first_party = "sym_cps"
include_trailing_comma = true
